+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- POST request for loading group members [#455](https://github.com/rokwire/groups-building-block/issues/455)
 
 ## [1.46.3] - 2024-07-24
 ### Fixed